requirements.txt
fastapi
uvicorn
pydantic
prometheus-fastapi-instrumentator
joblib
pandas
# Force NumPy 2 to match your local Global environment
numpy<2.0.0
scikit-learn
lightgbm
mlflow
pyarrow
dask
pandera
boto3
python-dotenv
evidently
s3fs
boto3
pyarrow
fastparquet
python-dotenv